# Backend/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Social Media API
  description: A social media API including user, post and comment resources.
  contact:
    name: Asim
    email: adalata@uni.coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://ecology-fiction-3000.codio-box.uk/api/v1'
    description: Development server
  - url: 'localhost:3000/api/v1'
    description: Local server
...
paths:
...
  /posts/{id}:
    summary: Represents an individual post.
    description: >
      This resource represents an individual post in the system.
      Each post is identified by a numeric `id`.
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the post
    get:
      tags:
        - Posts
      description:
        Get post by id.
      responses:
        '200':
          description: View of a post.
        '404':
          description: Post was not found in the DB.
    put:
      tags:
        - Posts
      description:
        Update post by ID.
      requestBody:
        description: Existing post data, including edited data
        required: true
        content:
          application/json:
            schema:
              $ref: ./post.json#/definitions/post
      responses:
        '200':
          description: Update confirmation and link to record.
          content:
            application/json:
              schema:
                $ref: ./post.json#/definitions/postCreated
        '400':
          description: Request data not as expected.
        '401':
          description: Only authenticated users can update posts.
        '403':
          description: Only the owner of a post or an admin user can update it.
        '404':
          description: Post was not found in the DB.
    delete:
      tags:
        - Posts
      description:
        Delete post by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only authenticated users can delete posts.
        '403':
          description: Only the owner of a post or an admin user can delete it.
        '404':
          description: Post was not found in the DB.
  /posts:
    summary: Represents the post resource.
    description: >
      This resource represents the system's post resource.
    post:
      tags:
        - Posts
      description:
        Create a new post.
      requestBody:
        description: All new post data
        required: true
        content:
          application/json:
            schema:
              $ref: ./post.json#/definitions/post
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: ./post.json#/definitions/postCreated
        '400':
          description: Request data not as expected.
        '401':
          description: Only authenticated users can create posts.
    get:
      tags:
        - Posts
      description:
        Retrieve all posts.
      responses:
        '200':
          description: All posts.
        '404':
          description: No posts found in the DB.
  /posts/{postID}/comments/{id}:
    summary: Represents an individual comment.
    description: >
      This resource represents an individual comment in the system.
      The associated post is identified by a numeric `postID`.
      Each comment is identified by a numeric `id`.
    parameters:
        - in: path
          name: postID
          schema:
            type: integer
          required: true
          description: Numeric id of the associated post
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the comment
    get:
      tags:
        - Comments
      description:
        Get comment by id.
      responses:
        '200':
          description: View of a comment.
        '404':
          description: Comment was not found in the DB.
    put:
      tags:
        - Comments
      description:
        Update comment by ID.
      requestBody:
        description: Existing comment data, including edited data
        required: true
        content:
          application/json:
            schema:
              $ref: ./comment.json#/definitions/comment
      responses:
        '200':
          description: Update confirmation and link to record.
          content:
            application/json:
              schema:
                $ref: ./comment.json#/definitions/commentCreated
        '400':
          description: Request data not as expected.
        '401':
          description: Only authenticated users can update comments.
        '403':
          description: Only the owner of a comment or an admin user can update it.
        '404':
          description: Comment was not found in the DB.
    delete:
      tags:
        - Comments
      description:
        Delete comment by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only authenticated users can delete posts.
        '403':
          description: Only the owner of a comment or an admin user can delete it.
        '404':
          description: Comment was not found in the DB.
  /posts/{postID}/comments:
    summary: Represents the comment resource.
    description: >
      This resource represents the system's comment resource.
      The associated post is identified by a numeric `postID`.
    parameters:
        - in: path
          name: postID
          schema:
            type: integer
          required: true
          description: Numeric id of the associated post
    post:
      tags:
        - Comments
      description:
        Create a new comment.
      requestBody:
        description: All new comment data
        required: true
        content:
          application/json:
            schema:
              $ref: ./comment.json#/definitions/comment
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: ./comment.json#/definitions/commentCreated
        '400':
          description: Request data not as expected.
        '401':
          description: Only authenticated users can create comments.
    get:
      tags:
        - Comments
      description:
        Retrieve all comments.
      responses:
        '200':
          description: All comments.
        '404':
          description: No comments found in the DB.
...